refactor : css 리팩토링 및 폴더 구조 변경
