+++ src/App.jsx
@@ -1,14 +1,14 @@
-import Input from './components/input';
-import DropDown from './components/dropdown';
-import ReactDraft from './components/reactDraft/index';
+import Input from './components/Input';
+import DropDown from './components/Dropdown/Dropdown';
+import ReactDraft from './components/ReactDraft/ReactDraft';
 
 function App() {
-  let a = [123, 123, 234, 345];
+  //let a = [123, 123, 234, 345];
   return (
     <>
-      <Input placeholder="플레이스호더" errorMassage="에러다"></Input>
+      {/* <Input placeholder="플레이스호더" errorMassage="에러다"></Input>
       <DropDown label="친구" name="친구" placeholders={a} />
-      <ReactDraft />
+      <ReactDraft /> */}
     </>
   );
 }
