+++ src/components/Input/Input.module.scss
@@ -1,23 +1,23 @@
-.input--inactive {
+.inputInactive {
   border: 1px solid $color-gray-300;
 }
 
-.input--active {
+.inputActive {
   border: 2px solid $color-gray-700;
 }
 
-.input--focused {
+.inputFocused {
   border: 2px solid $color-gray-700;
 }
 
-.input--defalut:hover {
+.inputDefalut:hover {
   border: 1px solid $color-gray-500;
 }
 
-.input--error {
+.inputError {
   border: 1px solid $color-error;
 }
-.input--defalut {
+.inputDefalut {
   display: flex;
   width: min(720px, 100% - 48px);
   padding: 12px 16px;
@@ -27,11 +27,11 @@
   border-radius: 8px;
 }
 
-.p--error {
+.pError {
   color: $color-error;
   @include font-12-Regular;
   display: inline-block;
 }
-.p--default {
+.pDefault {
   display: none;
 }
