+++ src/components/Input/index.js
@@ -1,0 +1,1 @@
+export { default as Input } from './Input';
