+++ src/components/ReactDraft/ReactDraft.jsx
@@ -1,8 +1,8 @@
-import { convertToRaw, EditorState } from 'draft-js';
+import { EditorState } from 'draft-js';
 import { useState } from 'react';
 import { Editor } from 'react-draft-wysiwyg';
 import 'react-draft-wysiwyg/dist/react-draft-wysiwyg.css';
-import './styles.scss';
+import './ReactDraft.module.scss';
 
 export default function Draft() {
   const [editorState, setEditorState] = useState(EditorState.createEmpty());
