+++ src/hooks/useManageDropdown/useManageDropdown.js
@@ -1,13 +1,11 @@
 import { useState, useRef, useEffect } from 'react';
+import styles from '../../components/Dropdown/Dropdown.module.scss';
 
 const useManageDropdown = (placeholders) => {
   const [isUlClicked, setIsUlClicked] = useState(false); //ul 클릭 -> 자동으로 blur됨.
   const [isLiClicked, setIsLiClicked] = useState(false);
   const [clickedLi, setClickedLi] = useState(placeholders[0]);
-  //active : ul한번 클릭돼서 li 열렸을 때
-  // focused:콜라서 닫힘
-  // hovered: hover
-  //아무것도 아닐 때 : inactive
+
   const dropDownRef = useRef(null);
 
   useEffect(() => {
@@ -15,7 +13,6 @@ const useManageDropdown = (placeholders) => {
       if (dropDownRef.current && !dropDownRef.current.contains(e.target)) {
         setIsUlClicked(false);
         setIsLiClicked(false);
-        console.log('handleClickoutSide 동작', isUlClicked);
       }
     };
     document.addEventListener('mousedown', handleClickOutside);
@@ -34,9 +31,9 @@ const useManageDropdown = (placeholders) => {
   };
 
   const decideSelectClass = () => {
-    if (isUlClicked && !isLiClicked) return 'select--active';
-    else if (isLiClicked) return 'select--focused';
-    else return 'select--inactive';
+    if (isUlClicked && !isLiClicked) return styles.selectActive;
+    else if (isLiClicked) return styles.selectFocused;
+    else return styles.selectInactive;
   };
 
   return {
